// src/types/generated-client/models/NotificationList.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 같이 달램 API
 * 모임 매칭 플랫폼 API  ## 인증 방식 - **JWT Bearer Token**: 로그인 후 발급받은 accessToken을 Authorization 헤더에 포함 - accessToken 만료 시간: 15분 - refreshToken 만료 시간: 7일  ## 사용 예시 ``` Authorization: Bearer eyJhbGciOiJIUzI1NiIs... ```
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Notification } from './Notification';
import {
  NotificationFromJSON,
  NotificationFromJSONTyped,
  NotificationToJSON,
  NotificationToJSONTyped,
} from './Notification';

/**
 *
 * @export
 * @interface NotificationList
 */
export interface NotificationList {
  /**
   *
   * @type {Array<Notification>}
   * @memberof NotificationList
   */
  data: Array<Notification>;
  /**
   *
   * @type {string}
   * @memberof NotificationList
   */
  nextCursor: string;
  /**
   *
   * @type {boolean}
   * @memberof NotificationList
   */
  hasMore: boolean;
}

/**
 * Check if a given object implements the NotificationList interface.
 */
export function instanceOfNotificationList(value: object): value is NotificationList {
  if (!('data' in value) || value['data'] === undefined) return false;
  if (!('nextCursor' in value) || value['nextCursor'] === undefined) return false;
  if (!('hasMore' in value) || value['hasMore'] === undefined) return false;
  return true;
}

export function NotificationListFromJSON(json: any): NotificationList {
  return NotificationListFromJSONTyped(json, false);
}

export function NotificationListFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean
): NotificationList {
  if (json == null) {
    return json;
  }
  return {
    data: (json['data'] as Array<any>).map(NotificationFromJSON),
    nextCursor: json['nextCursor'],
    hasMore: json['hasMore'],
  };
}

export function NotificationListToJSON(json: any): NotificationList {
  return NotificationListToJSONTyped(json, false);
}

export function NotificationListToJSONTyped(
  value?: NotificationList | null,
  ignoreDiscriminator: boolean = false
): any {
  if (value == null) {
    return value;
  }

  return {
    data: (value['data'] as Array<any>).map(NotificationToJSON),
    nextCursor: value['nextCursor'],
    hasMore: value['hasMore'],
  };
}